Refactoring: add `from_smartdevice?`
