+++ lib/rack/simple_user_agent/detectors/windows_phone.rb
@@ -1,0 +1,11 @@
+module Rack
+  class SimpleUserAgent
+    module Detectors
+      module WindowsPhone
+        def from_windows_phone?
+          user_agent_string.include?("Windows Phone")
+        end
+      end
+    end
+  end
+end
